Another refactor...
