+++ superset-frontend/packages/superset-ui-core/src/query/index.ts
@@ -31,6 +31,7 @@ export { default as normalizeOrderBy } from './normalizeOrderBy';
 export { normalizeTimeColumn } from './normalizeTimeColumn';
 export { default as extractQueryFields } from './extractQueryFields';
 export * from './getXAxis';
+export * from './getClientErrorObject';
 
 export * from './types/AnnotationLayer';
 export * from './types/QueryFormData';
