+++ superset-frontend/packages/superset-ui-core/src/utils/index.ts
@@ -28,9 +28,7 @@ export { default as logging } from './logging';
 export { default as removeDuplicates } from './removeDuplicates';
 export { lruCache } from './lruCache';
 export { getSelectedText } from './getSelectedText';
-export { SEPARATOR, fetchTimeRange } from './fetchTimeRange';
 export * from './featureFlags';
 export * from './random';
 export * from './typedMemo';
 export * from './html';
-export * from './getClientErrorObject';
